Use built in rating bar.

+++ app/src/main/java/ru/sandbox/androidacademyapp/data/db/entities/Movie.kt
@@ -24,7 +24,7 @@ data class Movie(
     val backdropUrl: String,
 
     @ColumnInfo(name = Movies.COLUMN_NAME_RATINGS)
-    val ratings: Int,
+    val ratings: Float,
 
     @ColumnInfo(name = Movies.COLUMN_NAME_REVIEWS)
     val reviews: Int,
